v1.0.2: Set power_bool to 1 after pressing Up, Down or Reset
This fixes an issue where the power state boolean could stay as 0 even if the dimmer was on.
Usually this could happen because of powering off, then turning up/down/resetting the power and trying to power off again,
which would make the power button not work unless pressed two times.

Also I forgot to bump the project version in the file for 1.0.1 lol

Signed-off-by: Flopster101 <nahuelgomez329@gmail.com>